+++ pom.xml
@@ -12,7 +12,7 @@
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>4.12</version>
-			<scope>ppdd</scope>
+			<scope>test</scope>
 			<optional>true</optional>
 		</dependency>
 	</dependencies>
