+++ src/test/java/test/ClassIShouldFindOne.java
@@ -1,4 +1,4 @@
-package ppdd;
+package test;
 
 public class ClassIShouldFindOne {
 	public static void foo() {
