+++ src/test/java/test/ClassIShouldFindTwo.java
@@ -1,4 +1,4 @@
-package ppdd;
+package test;
 
 public class ClassIShouldFindTwo {
 	public static void foo() {
