+++ src/test/java/test/subpkg/ClassIShouldFindThree.java
@@ -1,4 +1,4 @@
-package ppdd.subpkg;
+package test.subpkg;
 
 public class ClassIShouldFindThree {
 	public static void foo() {
